pyproject.toml
[project]
name = "rishty-wali-aunty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.95.1",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
    "twilio>=9.6.5",
]
